Página de Perfil de usuario
Añadidas varias funcionalidades:
- Página de perfil para que los usuarios puedan ver sus películas pendientes y vistas.
- Las miniaturas de las películas ahora muestran su puntuación media, además de la puntuación del usuario logueado.
- Unificado código JS para tener mas código reutilizado.

+++ src/main/java/org/example/videoclub/models/dto/PeliculaCompletaDTO.java
@@ -14,7 +14,7 @@ public class PeliculaCompletaDTO {
     private Integer anio;
     private Integer duracion;
     private String sinopsis;
-    private BigDecimal puntuacionMedia;
+    private Integer puntuacionMedia;
     private Long numVotos;
 
     private Map<String, String> mapGeneros;
@@ -89,11 +89,11 @@ public class PeliculaCompletaDTO {
         this.sinopsis = sinopsis;
     }
 
-    public BigDecimal getPuntuacionMedia() {
+    public Integer getPuntuacionMedia() {
         return puntuacionMedia;
     }
 
-    public void setPuntuacionMedia(BigDecimal puntuacionMedia) {
+    public void setPuntuacionMedia(Integer puntuacionMedia) {
         this.puntuacionMedia = puntuacionMedia;
     }
